// src/app/entities/finalidad-interface.ts
export interface FinalidadInterface {
  id?: number;
  id_ciclo?: number;
  codigo?: string;
  descripcion?: string;
  id_funcion?: number;
  id_div_funcional?: number;
  id_gru_funcional?: number;
  id_unidad?: number;
  estado?: number;
  usuario_registro?: string;
  fecha_registro?: string;
  nombre_usuario?: string;
}